Adds in Windows Docker CLI, initial changes to volumes to support Windows
