+++ baseimages/docker-cli/docker.Dockerfile
