+++ builder/constants_windows.go
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package builder
+
+const (
+	// homeWorkDir is the working directory to start at in $HOME
+	homeWorkDir = "c:\\acb\\home"
+
+	// containerWorkspaceDir is the default working directory for a container.
+	containerWorkspaceDir = "c:\\workspace"
+)
